+++ CHANGELOG.md
@@ -1,13 +1,12 @@
 # 0.3.0
 + Add settings page for user
++ Retrieve user settings
 
 # 0.2.0
-## 29-03-2019
 + Add bootstrap layout
 + Add Public IP on dashboard
 + Enable change password
 
 # 0.1.0
-## 24-03-2019
 + First boot
 + Add get public IP